requirements.txt
albucore==0.0.13
albumentations==1.4.13
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
Babel==2.15.0
backcall==0.2.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.21.2
einops==0.8.0
eval_type_backport==0.2.0
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.13.1
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.2.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
imageio==2.34.2
importlib_metadata==8.0.0
intel-openmp==2021.4.0
ipykernel==6.29.0
ipython==8.12.3
ipywidgets==8.1.3
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
keyring==25.2.1
kiwisolver==1.4.5
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplot==0.1.9
matplotlib==3.9.1
matplotlib-inline==0.1.6
mdurl==0.1.2
MedPy==0.5.2
mistune==3.0.2
mkl==2021.4.0
monai==1.3.2
more-itertools==10.3.0
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nh3==0.2.17
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.3
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
pkginfo==1.10.0
platformdirs==4.1.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==5.27.3
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyloco==0.0.139
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
pywin32-ctypes==0.2.2
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.2
QtPy==2.4.1
readme_renderer==43.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.0
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.12.0
setproctitle==1.3.3
SimpleITK==2.3.1
SimpleWebSocketServer==0.1.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tbb==2021.11.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.7.24
timm==1.0.8
tinycss2==1.3.0
tomli==2.0.1
torch==2.3.1+cu118
torchaudio==2.3.1+cu118
torchvision==0.18.1+cu118
tornado==6.4
tqdm==4.66.5
traitlets==5.14.1
twine==5.1.1
types-python-dateutil==2.9.0.20240316
typing==3.7.4.3
typing_extensions==4.9.0
tzdata==2024.1
unzip==1.0.0
uri-template==1.3.0
urllib3==2.2.2
ushlex==0.99.1
uv==0.2.21
wandb==0.17.6
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.11
yarg==0.1.9
zipp==3.19.2
